get_form_django_drf/requirements.txt
Django==5
Pillow==9.5.0
python-dotenv==1.0.1
pymongo==4.10.1
djangorestframework==3.15.1